add open data query
