+++ src/index.d.ts
@@ -9,5 +9,6 @@ export * from './closures/PrototypeMethodParser';
 export * from './closures/MathMethodParser';
 export * from './closures/DateMethodParser';
 export * from './closures/StringMethodParser';
+export * from './open-data-query.expression';
 export * from './open-data-query.formatter';
