+++ src/index.js
@@ -10,4 +10,5 @@ export * from './closures/MathMethodParser';
 export * from './closures/DateMethodParser';
 export * from './closures/StringMethodParser';
 export * from './object-name.validator';
+export * from './open-data-query.expression';
 export * from './open-data-query.formatter';
