+++ src/open-data-query.expression.d.ts
@@ -1,0 +1,5 @@
+import { QueryExpression } from './query';
+
+export declare class OpenDataQuery extends QueryExpression {
+    //
+}
